fix: SFC transpilation in es buld (#1410)
* Create _form-file.js

* ESLint

* Update form-file.vue

* ESLint

* Create _form-group.js

* Update form-group.vue

* ESLint

* Create _form-input.js

* Update form-input.vue

* ESLint

* ESLint

* Create _progress.js

* Update progress.vue

* Create _modal.js

* Update modal.vue

* ESLint
